More Inventory and Inventory UI work
Finished up reworking the interaction with items that can be picked up to the player inventory and have started using an event handler in the UI to start updating the inventory UI when the player is interacting with items.  Right now the only event  is when an item is added to the inventory from the ground.