More work on inventory system
Can now click on items in inventory panel to equip / unequip. These button components call the equip and unequip functions of the inventoryController. Still need to fix a bug where you can click on equipped items as rendered on the playerModel and the player will "Take" them, creating a duplicate UI item for the equipped item in the player's bag which can throw an exception when clicked.