Working on a fix for Scaling issues with gameobjects that will be parent of equipment models
When I equipped the sword using the new player model I noticed it was scaled unintentionally. I've created a pattern for the playermodel hierarchy that will solve the problem and applied it to the right shoulder and all subsequent children. This solution still needs to be applied to the rest of the player model.