Quick n dirty fix for sword's orientation when equipped.
Fixed a bug where an item being equipped couldn't have it's rotation/position manipulated do to being under control of an active idle animation when the appropriate part of the player model was set as the item's parent. This implementation needs to be abstracted a bit and made to handle items starting from any orientation.