Created EnemyHealth script and set up new canvas
there is a new canvas object that's a child of the Enemies container object. The goal is to have an enemies created in game from a prefab and have an EnemyHealthBar created in the new canvas. The spawned enemy will have a reference to the healthbar created along side and update that ui element. Finally, when an enemy is killed, it's Die() function should destroy the healthbar object.

Currently there is a test Slime and EnemyHealthBar in the hierarchy for testing. The healthbar is not maintaining a position above the Slime object and that needs to be figured out before setting up the spawning system of the prefabs.