Projectiles fire again.
Since I now have the weapon specific scripts inherit from the new EquippableModel class that is attached to the model objects, I needed to change how the InventoryController handled the projectileSpawn property on equip to give that reference to the model instead of container object of equippables.

Also I added a mainHand reference to the PlayerAnimationController that grabs the appropriate object in the Start() function and lowers the expense of the other functions inside that class, as well as adding to the FireProjectile() function to actually invoke the weapon's FireProjectile().